Move Library DTOs to libcalibre root
