--- src-tauri/libcalibre/src/application/services/library/mod.rs
@@ -1,1 +1,0 @@
-pub mod dto;
