+++ src-tauri/libcalibre/src/application/services/mod.rs
@@ -1,2 +1,1 @@
 pub mod domain;
-pub mod library;
