+++ src-tauri/libcalibre/src/dtos/mod.rs
@@ -1,1 +1,2 @@
 pub mod author;
+pub mod library;
